Deleted group_3 and moved functions to group_4 and group_5
