+++ group_4.py
@@ -117,6 +117,25 @@ def tc_dna_intent_api_v1_interface_count():
     tc.okay('complete')
 
 
+#dna/intent/api/v1/network-device
+def tc_dna_intent_api_v1_network_device():
+    # create this test case
+    tc = TestCase(test_name='IntentApiV1NetworkDevice', yaml_file='params.yaml')
+
+    # create a session to the DNA-C
+    dnac = DnaCenter(hostname=tc.params['DnaCenter']['Hostname'],
+                     port=tc.params['DnaCenter']['Port'],
+                     username=tc.params['DnaCenter']['Username'],
+                     password=tc.params['DnaCenter']['Password'])
+
+    # execute the command and get response
+    response = dnac.get('dna/intent/api/v1/network-device')
+    pp.pprint(response.json())
+
+    # complete
+    tc.okay('complete')
+
+
 def run_all_tests():
     # run this test case first since it will do a basic 'ping'
     tc_dna_intent_api_v1_network_device_count()
@@ -125,5 +144,7 @@ def run_all_tests():
     tc_dna_intent_api_v1_interface()
     tc_dna_intent_api_v1_interface_network_device()
     tc_dna_intent_api_v1_interface_count()
+    tc_dna_intent_api_v1_network_device()
+
 
 run_all_tests()
